pyproject.toml
[project]
name = "hw-django-project"
version = "0.1.0"
description = ""
authors = [
    {name = "MaksimTishchenko",email = "tmaxim.1997@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
